Conflicts:
	src/voxeleffect/voxelparticleworld.cpp

+++ src/geometry/abstractshape.h
@@ -7,7 +7,7 @@ class WorldTransform;
 class AbstractShape {
 public:
     virtual bool intersects(const Sphere& sphere) const = 0;
-    virtual bool nearTo(const TAABB<float>& aabb) const = 0;
-    virtual bool containedBy(const TAABB<float>& aabb) const = 0;
+    virtual bool nearTo(const TAABB<int>& aabb) const = 0;
+    virtual bool containedBy(const TAABB<int>& aabb) const = 0;
 };
